+++ src/Console/Commands/ModelsCommand.php
@@ -67,7 +67,7 @@ class ModelsCommand extends Command
                 'concrete'  => $item
             ];
         })->sort(function ($a, $b) {
-            return $a['shortName'] > $b['shortName'];
+            return $a['shortName'] <=> $b['shortName'];
         })->each(function ($binding) use (&$table) {
             $table[] = [
                 $binding['shortName'],